Move AppPayPayButton component to matters dir
