+++ pages/index.vue
@@ -85,7 +85,7 @@ import AppGifteeButton from '~/components/matters/AppGifteeButton'
 import AppHeader1 from '~/components/elements/AppHeader1'
 import AppHeader2 from '~/components/elements/AppHeader2'
 import AppLinkButton from '~/components/elements/AppLinkButton'
-import AppPayPayButton from '~/components/AppPayPayButton'
+import AppPayPayButton from '~/components/matters/AppPayPayButton'
 import AppProfileIcon from '~/components/AppProfileIcon'
 import AppSkillButton from '~/components/AppSkillButton'
 import AppText from '~/components/elements/AppText'
